move data-dir to wanikani
