+++ level_progress/unlocks_test.go
@@ -2,7 +2,7 @@ package level_progress
 
 import (
 	"encoding/json"
-	"github.com/mdreem/wanikani_cli/data"
+	"github.com/mdreem/wanikani_cli/wanikani/data"
 	"reflect"
 	"testing"
 	"time"
