+++ progress_info/progress_info.go
@@ -2,8 +2,8 @@ package progress_info
 
 import (
 	"fmt"
-	"github.com/mdreem/wanikani_cli/data"
 	"github.com/mdreem/wanikani_cli/wanikani"
+	"github.com/mdreem/wanikani_cli/wanikani/data"
 	"github.com/spf13/cobra"
 )
