+++ progress_info/progress_info_test.go
@@ -2,7 +2,7 @@ package progress_info
 
 import (
 	"encoding/json"
-	"github.com/mdreem/wanikani_cli/data"
+	data2 "github.com/mdreem/wanikani_cli/wanikani/data"
 	"reflect"
 	"testing"
 )
@@ -35,11 +35,11 @@ func Test_fetchDistribution(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var assignments []data.AssignmentEnvelope
+			var assignments []data2.AssignmentEnvelope
 			for _, srsLevel := range tt.srsLevels {
 				assignments = append(assignments, createAssignment(srsLevel))
 			}
-			testClient := data.TestClient{
+			testClient := data2.TestClient{
 				Assignments: assignments,
 			}
 
@@ -50,9 +50,9 @@ func Test_fetchDistribution(t *testing.T) {
 	}
 }
 
-func createAssignment(srsLevel json.Number) data.AssignmentEnvelope {
-	return data.AssignmentEnvelope{
-		Data: data.Assignment{
+func createAssignment(srsLevel json.Number) data2.AssignmentEnvelope {
+	return data2.AssignmentEnvelope{
+		Data: data2.Assignment{
 			SrsStage: srsLevel,
 		},
 	}
@@ -107,11 +107,11 @@ func Test_ComputeProgressInfo(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var assignments []data.AssignmentEnvelope
+			var assignments []data2.AssignmentEnvelope
 			for _, srsLevel := range tt.srsLevels {
 				assignments = append(assignments, createAssignment(srsLevel))
 			}
-			testClient := data.TestClient{
+			testClient := data2.TestClient{
 				Assignments: assignments,
 			}
